+++ docs/examples/docs-demo-six-theme-switcher-3.vue
@@ -1,0 +1,18 @@
+<template>
+<div>
+
+        <div style="background: grey">
+          <six-theme-switcher label="Switch between light and dark mode"></six-theme-switcher>
+        </div>
+      
+</div>
+</template>
+<style>
+
+</style>
+<script>
+export default {
+  name: 'docs-demo-six-theme-switcher-3',
+  mounted() {  }
+}
+</script>